+++ MaskClicker/Assets/Scripts/SaveGame.cs
@@ -5,8 +5,8 @@ using UnityEngine.UI;
 
 public class SaveGame : MonoBehaviour
 {
-    public int saveGameMask;
-    public static int saveValue = 50;
+    public float saveGameMask;
+    public static float saveValue = 50;
     public GameObject saveButton;
     public GameObject saveText;
 
@@ -27,9 +27,9 @@ public class SaveGame : MonoBehaviour
     public void SaveTheGame()
     {
         GlobalMasks.MaskCount -= saveValue;
-        PlayerPrefs.SetInt("SavedMasks", GlobalMasks.MaskCount);
-        PlayerPrefs.SetInt("SavedUpgrade", GlobalUpgrade.upgradPerSec);
+        PlayerPrefs.SetFloat("SavedMasks", GlobalMasks.MaskCount);
+        PlayerPrefs.SetFloat("SavedUpgrade", GlobalUpgrade.upgradPerSec);
         saveValue *= 2;
-        PlayerPrefs.SetInt("SavedValue", saveValue);
+        PlayerPrefs.SetFloat("SavedValue", saveValue);
     }
 }
